Partially move paint into many files
