+++ main/modes/paint/paint_ui.h
@@ -1,0 +1,4 @@
+#ifndef _PAINT_UI_H_
+#define _PAINT_UI_H_
+
+#endif
