+++ main/modes/paint/paint_util.h
@@ -1,0 +1,9 @@
+#ifndef _PAINT_UTIL_H_
+#define _PAINT_UTIL_H_
+
+#include "palette.h"
+#include "swadge_util.h"
+
+paletteColor_t getContrastingColor(paletteColor_t col);
+
+#endif
